requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0  # Alternative si besoin

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0  # Augmentation avancée

# Data Science Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=0.2.1  # Pour export Plotly

# Interprétabilité
shap>=0.42.0
grad-cam>=1.4.0
pytorch-grad-cam>=1.4.0

# Medical Imaging
openslide-python>=1.2.0  # Pour WSI
scikit-image>=0.21.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.7.0
tensorboard>=2.13.0

# Data Versioning
dvc>=3.0.0
dvc-gdrive>=2.20.0  # Si utilisation Google Drive

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0  # Configuration avancée

# Utilities
tqdm>=4.65.0
click>=8.1.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Medical/Bio specific
SimpleITK>=2.2.0  # Image processing médical
pydicom>=2.4.0  # Si DICOM files

# Performance
numba>=0.57.0  # Accélération calculs

# Autres
requests>=2.31.0
python-dateutil>=2.8.2
